+++ example/example_Nova_Animal.c
@@ -70,12 +70,12 @@ void example_Nova_Animal_Nova_destroy(example_Nova_Animal** this, nova_standard_
 
 int example_Nova_Animal_0_Nova_getNumLegs(example_Nova_Animal* this, nova_standard_exception_Nova_ExceptionData* exceptionData)
 {
-	return 0;
+	return (int)0;
 }
 
 int example_Nova_Animal_0_Nova_getNumEyes(example_Nova_Animal* this, nova_standard_exception_Nova_ExceptionData* exceptionData)
 {
-	return 0;
+	return (int)0;
 }
 
 nova_standard_Nova_String* example_Nova_Animal_0_Nova_getDescription(example_Nova_Animal* this, nova_standard_exception_Nova_ExceptionData* exceptionData)
