+++ example/example_Nova_HashMapDemo.c
@@ -72,12 +72,12 @@ void example_Nova_HashMapDemo_Nova_main(example_Nova_HashMapDemo* this, nova_sta
 	
 	l1_Nova_words = nova_standard_datastruct_Nova_HashMap_0_Nova_construct(0, exceptionData);
 	l1_Nova_nums = (int*)NOVA_MALLOC(sizeof(nova_standard_primitive_number_Nova_Int) * 6);
-	l1_Nova_nums[0] = 0;
-	l1_Nova_nums[1] = 1;
-	l1_Nova_nums[2] = 2;
-	l1_Nova_nums[3] = 3;
-	l1_Nova_nums[4] = 4;
-	l1_Nova_nums[5] = 5;
+	l1_Nova_nums[0] = (int)(0);
+	l1_Nova_nums[1] = (int)(1);
+	l1_Nova_nums[2] = (int)(2);
+	l1_Nova_nums[3] = (int)(3);
+	l1_Nova_nums[4] = (int)(4);
+	l1_Nova_nums[5] = (int)(5);
 	nova_standard_datastruct_Nova_HashMap_virtual1_Nova_put((nova_standard_datastruct_Nova_HashMap*)(l1_Nova_words), exceptionData, (nova_standard_Nova_Object*)(nova_standard_primitive_number_Nova_Int_Nova_construct(0, exceptionData, l1_Nova_nums[0])), (nova_standard_Nova_Object*)(nova_standard_Nova_String_2_Nova_construct(0, exceptionData, "Zero")));
 	nova_standard_datastruct_Nova_HashMap_virtual1_Nova_put((nova_standard_datastruct_Nova_HashMap*)(l1_Nova_words), exceptionData, (nova_standard_Nova_Object*)(nova_standard_primitive_number_Nova_Int_Nova_construct(0, exceptionData, l1_Nova_nums[1])), (nova_standard_Nova_Object*)(nova_standard_Nova_String_2_Nova_construct(0, exceptionData, "One")));
 	nova_standard_datastruct_Nova_HashMap_virtual1_Nova_put((nova_standard_datastruct_Nova_HashMap*)(l1_Nova_words), exceptionData, (nova_standard_Nova_Object*)(nova_standard_primitive_number_Nova_Int_Nova_construct(0, exceptionData, l1_Nova_nums[2])), (nova_standard_Nova_Object*)(nova_standard_Nova_String_2_Nova_construct(0, exceptionData, "Two")));
