+++ example/network/example_network_Nova_ClientDemo.c
@@ -74,7 +74,7 @@ void example_network_Nova_ClientDemo_Nova_main(example_network_Nova_ClientDemo* 
 	
 	l1_Nova_socket = nova_standard_network_Nova_ClientSocket_Nova_construct(0, exceptionData);
 	l1_Nova_ip = nova_standard_Nova_String_2_Nova_construct(0, exceptionData, "127.0.0.1");
-	l1_Nova_port = 5675;
+	l1_Nova_port = (int)(5675);
 	nova_standard_io_Nova_Console_1_Nova_writeLine(0, exceptionData, nova_standard_Nova_String_0_Nova_concat(nova_standard_Nova_String_2_Nova_construct(0, exceptionData, "Attempting to connect to "), exceptionData, nova_standard_Nova_String_virtual0_Nova_concat((nova_standard_Nova_String*)(l1_Nova_ip), exceptionData, nova_standard_Nova_String_0_Nova_concat(nova_standard_Nova_String_2_Nova_construct(0, exceptionData, ":"), exceptionData, nova_standard_primitive_number_Nova_Int_2_Nova_toString(0, exceptionData, l1_Nova_port)))));
 	l1_Nova_connected = nova_standard_network_Nova_ClientSocket_Nova_connect(l1_Nova_socket, exceptionData, l1_Nova_ip, l1_Nova_port);
 	if (l1_Nova_connected)
