+++ nova/standard/database/nova_standard_database_Nova_DBConnector.c
@@ -131,12 +131,12 @@ nova_standard_database_Nova_ResultSet* nova_standard_database_Nova_DBConnector_N
 		l2_Nova_numRows = (int)nova_num_rows(this->prv->nova_standard_database_Nova_DBConnector_Nova_mysql);
 		l2_Nova_numCols = (int)nova_num_cols(this->prv->nova_standard_database_Nova_DBConnector_Nova_result);
 		l2_Nova_rows = (nova_standard_Nova_String***)nova_gen_array(NOVA_MALLOC(sizeof(nova_standard_Nova_String) * l2_Nova_numRows * l2_Nova_numCols), (int[]) { l2_Nova_numRows, l2_Nova_numCols }, 0, 1, sizeof(nova_standard_Nova_String));
-		l4_Nova_i = 0;
+		l4_Nova_i = (int)(0);
 		for (; l4_Nova_i < l2_Nova_numRows; l4_Nova_i++)
 		{
 			int l6_Nova_j;
 			
-			l6_Nova_j = 0;
+			l6_Nova_j = (int)(0);
 			for (; l6_Nova_j < l2_Nova_numCols; l6_Nova_j++)
 			{
 				l2_Nova_rows[l4_Nova_i][l6_Nova_j] = nova_standard_Nova_String_2_Nova_construct(0, exceptionData, l1_Nova_arrays[l4_Nova_i][l6_Nova_j]);
