+++ nova/standard/exception/nova_standard_exception_Nova_ExceptionData.c
@@ -103,7 +103,7 @@ nova_standard_exception_Nova_ExceptionData* nova_standard_exception_Nova_Excepti
 		int l3_Nova_i;
 		
 		l1_Nova_list = l1_Nova_data->nova_standard_exception_Nova_ExceptionData_Nova_codes;
-		l3_Nova_i = 0;
+		l3_Nova_i = (int)(0);
 		for (; l3_Nova_i < l1_Nova_list->nova_standard_datastruct_list_Nova_Array_Nova_size; l3_Nova_i++)
 		{
 			if ((int)(((nova_standard_primitive_number_Nova_Int*)nova_standard_datastruct_list_Nova_Array_Nova_get(l1_Nova_list, exceptionData, l3_Nova_i))->nova_standard_primitive_number_Nova_Int_Nova_value) == nova_standard_exception_Nova_ExceptionData_Nova_code)
@@ -140,7 +140,7 @@ void nova_standard_exception_Nova_ExceptionData_Nova_jumpToBuffer(nova_standard_
 	l1_Nova_data = nova_standard_exception_Nova_ExceptionData_Nova_getCorrectData(this, exceptionData, nova_standard_exception_Nova_ExceptionData_Nova_code);
 	if (nova_standard_exception_Nova_ExceptionData_Nova_getParent(l1_Nova_data, exceptionData) == (nova_standard_exception_Nova_ExceptionData*)nova_null)
 	{
-		nova_standard_exception_Nova_ExceptionData_Nova_code = 1;
+		nova_standard_exception_Nova_ExceptionData_Nova_code = (int)(1);
 	}
 	l1_Nova_b = nova_standard_exception_Nova_ExceptionData_Nova_getBuffer(l1_Nova_data, exceptionData);
 	jump(*l1_Nova_b, nova_standard_exception_Nova_ExceptionData_Nova_code);
